fix : 환경 통합 및 버그 react persist 이슈 픽스
