+++ src/components/provider/Providers.tsx
@@ -1,18 +1,16 @@
-'use client'
-
 // import StoreProvider from '@/components/provider/StoreProvider'
-import { SessionProvider } from 'next-auth/react'
 import React from 'react'
 
 import MSWComponent from '@/components/mock-service/MSWComponent'
-import ReduxProvider from '@/store/provider'
+import AuthProvider from '@/components/provider/AuthProvider'
+import ReduxProvider from '@/components/provider/ReduxProvider'
 
 export default function Providers({ children }: { children: React.ReactNode }) {
   return (
     <MSWComponent>
-      <ReduxProvider>
-        <SessionProvider>{children}</SessionProvider>
-      </ReduxProvider>
+      <AuthProvider>
+        <ReduxProvider>{children}</ReduxProvider>
+      </AuthProvider>
     </MSWComponent>
   )
 }
