+++ src/components/provider/ReduxProvider.tsx
@@ -3,10 +3,10 @@ import { Provider } from 'react-redux'
 import { persistStore } from 'redux-persist'
 import { PersistGate } from 'redux-persist/integration/react'
 
-import store from './store'
+import store from '@/store/store'
 export let persistor = persistStore(store)
 
-export default function ReduxProvider({ children }) {
+export default function ReduxProvider({ children }: { children: React.ReactNode }) {
   return (
     <Provider store={store}>
       <PersistGate loading={null} persistor={persistor}>
