+++ src/components/provider/StoreProvider.tsx
@@ -3,7 +3,7 @@
 import React from 'react'
 import { Provider } from 'react-redux'
 
-import { store } from '@/store/store'
+import store from '@/store/store'
 
 export default function StoreProvider({ children }: { children: React.ReactNode }) {
   return <Provider store={store}>{children}</Provider>
